refactor: remove "gold standard" model
remove support for cpp implementation of branch predictor to test bsv predictor with. still need to rename a lot of stuff for it to make more sense
