+++ branch/gold_standard/Makefile
@@ -36,9 +36,7 @@ TOP_BSV_PREDICTOR_FILE ?= $(TOPDIR)/$(PRED_DIR)/$(PREDICTOR)/Bluespec/$(PREDICTO
 DIR_PREDICTOR_INTERFACE_LOCATION ?= $(BSV_INCLUDE)/Predictor.bsv
 
 
-GOLD_STANDARD_HEADER ?= $(INCLUDE)/gold_standard.hpp
 TOP_FRAMEWORK_FILE ?= $(TOPDIR)/gold_standard.cc
-INCLUDE_GOLD_STANDARD ?= $(PRED_DIR)/$(PREDICTOR)/Model/$(PREDICTOR).hpp
 
 BSIM_DIRS = -simdir $(SIM_DIR) -bdir $(SIM_DIR) -info-dir $(SIM_DIR)
 CFILE = $(BSV_SRC)/reciever.c
@@ -49,13 +47,8 @@ failure_check:
 	make all_bsim PREDICTOR=$(PREDICTOR) || make unset PREDICTOR=$(PREDICTOR)
 
 
-ifdef MODEL_OFF
 .PHONY: all_bsim
-all_bsim: full_clean turn_model_checking_off set_bsv_predictor compile  link unset
-else
-.PHONY: all_bsim
-all_bsim: full_clean  set_bsv_predictor compile  link unset set_gold_standard
-endif
+all_bsim: full_clean set_bsv_predictor compile link unset
 
 build_bsim:
 	mkdir -p $(BUILD_DIR)
@@ -73,15 +66,6 @@ link:
 	bsc  -I $(INCLUDE) -e $(TOPMODULE) -sim -o $(BSIM_EXE) $(BSIM_DIRS) $(BSC_LINK_FLAGS) $(BSC_PATHS) \
 		$(CFILE)
 	@echo Linking for Bluesim finished
-
-.PHONY: set_gold_standard
-set_gold_standard:
-	test -f $(TOP_FRAMEWORK_FILE) && test -f $(INCLUDE_GOLD_STANDARD) && \
-		sed -i '1s|.*|#include "$(INCLUDE_GOLD_STANDARD)"|' $(TOP_FRAMEWORK_FILE)
-
-turn_model_checking_off:
-	test -f $(TOP_FRAMEWORK_FILE) &&\
-		sed -i '1s|.*|#define MODEL_OFF|' $(TOP_FRAMEWORK_FILE)
 
 .PHONY: set_bsv_predictor
 set_bsv_predictor:
